vernieuwingen

+++ website/app/level/finance/facturen.php
@@ -8,7 +8,8 @@ require $rootlink. '/app/templates/header.php';
 
 <style>
     .rood {
-     background-color: red !important;
+        background-color: indianred !important;
+        color: lightgray !important;
 }
 
 </style>
@@ -34,31 +35,72 @@ require $rootlink. '/app/templates/header.php';
         </thead>
         <tbody>
         <?php
-
+        if(isset($_GET['id']))
+        {
         $sql = "SELECT * FROM factuur WHERE klant_nr = '".Security($_GET['id'])."'";
+        }
+        else
+        {
+         $sql = "SELECT * FROM factuur";
+      
+
+        }
         if (! $query = mysqli_query($con, $sql)){
             echo "Kan gegevens niet uit database halen";
         }
-        $sql = "SELECT factuur_tot FROM factuur WHERE DATE > factuur_tot";
+      
         if (mysqli_num_rows($query) > 0 ){
             while ($row = mysqli_fetch_object($query)){
                 echo "<tr>";
+
+        
+          if($row->factuur_tot < time())
+                {
+                echo "<td class='rood'>" . $row->klant_nr . "</td>";
+                echo "<td class='rood'>" . $row->factuur_nr . "</td>";
+                echo "<td class='rood'>" . $row->bedrag . "</td>";
+                echo "<td class='rood'>" . $row->project_nr . "</td>";
+                echo "<td class='rood'>" . $row->btw . "</td>";
+                echo "<td class='rood'>" . $row->factuur_begin . "</td>";
+                echo "<td class='rood'>" . $row->factuur_tot . "</td>";
+                echo "<td class='rood'>" . $row->hoeveelheid . "</td>";
+                echo "<td class='rood'>" . $row->beschrijving . "</td>";
+                echo "<td class='rood'>" . $row->aantal . "</td>";
+                echo "<td class='rood'>" . $row->status . "</td>";
+                // echo "<td class='rood'><a href='edit.php?id=". $row->klant_nr . "'> Bewerk </a></td>";
+                // echo "<td class='rood'><a href='delete.php?id=" . $row->klant_nr . "'> X </a></td>";
+                }
+                else
+                {
+              
+                
+
+
                 echo "<td>" . $row->klant_nr . "</td>";
                 echo "<td>" . $row->factuur_nr . "</td>";
                 echo "<td>" . $row->bedrag . "</td>";
                 echo "<td>" . $row->project_nr . "</td>";
                 echo "<td>" . $row->btw . "</td>";
-                echo "<td>" . $row->factuur_tot . "</td>";
                 echo "<td>" . $row->factuur_begin . "</td>";
+                echo "<td>" . $row->factuur_tot . "</td>";
                 echo "<td>" . $row->hoeveelheid . "</td>";
                 echo "<td>" . $row->beschrijving . "</td>";
                 echo "<td>" . $row->aantal . "</td>";
                 echo "<td>" . $row->status . "</td>";
                 echo "<td><a href='edit.php?id=". $row->klant_nr . "'> Bewerk </a></td>";
                 echo "<td><a href='delete.php?id=" . $row->klant_nr . "'> X </a></td>";
+
+                }
+
+
                 echo "</tr>";
             }
+
+         
         }
+
+
+
